Correcoes

+++ cuda/main.c
@@ -40,8 +40,8 @@ int main (int argc, char **argv){
 
     // DEFINE A QUANTIDADE DE LINHAS
     // DA IMAGEM PARA LEITURA E SMOOTH
-    int numMaxGrids = (65535 / (imageParams->linha * imageParams->coluna));
-        printf("\nE %d", numMaxGrids);
+    float numMaxGrids = (65535 / (imageParams->linha * imageParams->coluna));
+        printf("\nE %f", numMaxGrids);
     exit(0);
     int numMaxLinhas = imageParams->linha;
     if (numMaxGrids > 1)
